Move form types into Form/Type folder
