+++ src/AppBundle/Form/Type/AdventureType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use AppBundle\Entity\Adventure;
 use AppBundle\Entity\Author;
