+++ src/AppBundle/Form/Type/AuthorType.php
@@ -1,15 +1,14 @@
 <?php
 
+namespace AppBundle\Form\Type;
 
-namespace AppBundle\Form;
-
-use AppBundle\Entity\Item;
+use AppBundle\Entity\Author;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class ItemType extends AbstractType
+class AuthorType extends AbstractType
 {
     /**
      * {@inheritdoc}
@@ -20,14 +19,14 @@ class ItemType extends AbstractType
             'required' => true,
         ]);
     }
-
+    
     /**
      * {@inheritdoc}
      */
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => Item::class
+            'data_class' => Author::class
         ]);
     }
 }
