+++ src/AppBundle/Form/Type/BossMonsterType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use AppBundle\Entity\Monster;
 use AppBundle\Entity\MonsterType as MonsterTypeEntity;
@@ -8,9 +8,11 @@ use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\FormEvent;
+use Symfony\Component\Form\FormEvents;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class CommonMonsterType extends AbstractType
+class BossMonsterType extends AbstractType
 {
     /**
      * {@inheritdoc}
@@ -27,7 +29,14 @@ class CommonMonsterType extends AbstractType
                 'required' => true,
                 'class' => MonsterTypeEntity::class,
                 'multiple' => true,
-            ]);
+            ])
+        ->addEventListener(FormEvents::POST_SUBMIT, function (FormEvent $formEvent) {
+            /** @var Monster $monster */
+            $monster = $formEvent->getData();
+            if ($monster !== null) {
+                $monster->setIsUnique(true);
+            }
+        });
     }
     
     /**
