+++ src/AppBundle/Form/Type/ChangeRequestType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use AppBundle\Entity\ChangeRequest;
 use AppBundle\Field\FieldProvider;
