+++ src/AppBundle/Form/Type/ItemType.php
@@ -1,14 +1,15 @@
 <?php
 
-namespace AppBundle\Form;
 
-use AppBundle\Entity\Author;
+namespace AppBundle\Form\Type;
+
+use AppBundle\Entity\Item;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class AuthorType extends AbstractType
+class ItemType extends AbstractType
 {
     /**
      * {@inheritdoc}
@@ -19,14 +20,14 @@ class AuthorType extends AbstractType
             'required' => true,
         ]);
     }
-    
+
     /**
      * {@inheritdoc}
      */
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => Author::class
+            'data_class' => Item::class
         ]);
     }
 }
